fix(ui): keep overview banner variants server-safe
