+++ ui/app/(prowler)/_overview/_lib/overview-banner.ts
@@ -1,0 +1,7 @@
+export const OVERVIEW_BANNER_VARIANT = {
+  LIGHTHOUSE: "lighthouse",
+  AGENTS: "agents",
+} as const;
+
+export type OverviewBannerVariant =
+  (typeof OVERVIEW_BANNER_VARIANT)[keyof typeof OVERVIEW_BANNER_VARIANT];
